Add Rstudio R-markdown code.
R studio will generate HTML with graphs from ".Rmd" files,
which contain special, markdown for R code.

The Rproj file configures Rstudio projects.
The .Rmd file is just a link to the .md file, but displays correctly in Rstudio,
while displaying as well as can be expected on GitHub itself.